+++ src/clflags.ml
@@ -10,3 +10,4 @@ let workspace_root = ref "."
 let external_lib_deps_hint = ref []
 let capture_outputs = ref true
 let debug_backtraces = ref false
+let diff_command = ref None