+++ src/diff.mli
@@ -1,0 +1,2 @@
+(** Diff two files that are expected not to match. *)
+val print : Path.t -> Path.t -> _ Future.t