+++ src/jbuild_load.ml
@@ -145,7 +145,7 @@ end
                Did you forgot to call [Jbuild_plugin.V*.send]?"
             (Path.to_string file);
         let sexps = Sexp_lexer.Load.many (Path.to_string generated_jbuild) in
-        return (dir, scope, Stanzas.parse scope sexps))
+        return (dir, scope, Stanzas.parse ~syntax:OCaml scope sexps))
     |> Future.all
 end
 
@@ -159,7 +159,7 @@ let load ~dir ~scope =
   let file = Path.relative dir "jbuild" in
   match Sexp_lexer.Load.many_or_ocaml_script (Path.to_string file) with
   | Sexps sexps ->
-    Jbuilds.Literal (dir, scope, Stanzas.parse scope sexps)
+    Jbuilds.Literal (dir, scope, Stanzas.parse scope ~syntax:Plain sexps)
   | Ocaml_script ->
     Script { dir; scope }